Move a few test functions related to auth tokens to separate file (#9018)
For readability. neon_fixtures.py is huge.
